requirements.txt
mail-parser
python-dotenv
fastapi
uvicorn[standard]
